+++ packages/util-stream/src/checksum/createChecksumStream.browser.ts
@@ -1,36 +1,10 @@
-import { Checksum, Encoder } from "@smithy/types";
 import { toBase64 } from "@smithy/util-base64";
 
 import { isReadableStream } from "../stream-type-check";
+import { ChecksumStream, ChecksumStreamInit } from "./ChecksumStream.browser";
 
 /**
  * @internal
- */
-export interface ChecksumStreamInit {
-  /**
-   * Base64 value of the expected checksum.
-   */
-  expectedChecksum: string;
-  /**
-   * For error messaging, the location from which the checksum value was read.
-   */
-  checksumSourceLocation: string;
-  /**
-   * The checksum calculator.
-   */
-  checksum: Checksum;
-  /**
-   * The stream to be checked.
-   */
-  source: ReadableStream;
-
-  /**
-   * Optional base 64 encoder if calling from a request context.
-   */
-  base64Encoder?: Encoder;
-}
-
-/**
  * Alias prevents compiler from turning
  * ReadableStream into ReadableStream<any>, which is incompatible
  * with the NodeJS.ReadableStream global type.
@@ -106,14 +80,3 @@ export const createChecksumStream = ({
   Object.setPrototypeOf(readable, ChecksumStream.prototype);
   return readable;
 };
-
-const ReadableStreamRef = typeof ReadableStream === "function" ? ReadableStream : function (): void {};
-
-/**
- * This stub exists so that the readable returned by createChecksumStream
- * identifies as "ChecksumStream" in alignment with the Node.js
- * implementation.
- *
- * @extends ReadableStream
- */
-export class ChecksumStream extends (ReadableStreamRef as any) {}
